+++ frontend/src/components/prompts/Share.vue
@@ -3,6 +3,7 @@
     <div class="card-title">
       <h2>{{ $t("buttons.share") }}</h2>
     </div>
+    <div class="searchContext">Path: {{ getContext }}</div>
 
     <template v-if="listing">
       <div class="card-content">
@@ -17,9 +18,7 @@
           <tr v-for="link in links" :key="link.hash">
             <td>{{ link.hash }}</td>
             <td>
-              <template v-if="link.expire !== 0">{{
-                humanTime(link.expire)
-              }}</template>
+              <template v-if="link.expire !== 0">{{ humanTime(link.expire) }}</template>
               <template v-else>{{ $t("permanent") }}</template>
             </td>
             <td class="small">
@@ -96,11 +95,7 @@
           </select>
         </div>
         <p>{{ $t("prompts.optionalPassword") }}</p>
-        <input
-          class="input input--block"
-          type="password"
-          v-model.trim="password"
-        />
+        <input class="input input--block" type="password" v-model.trim="password" />
       </div>
 
       <div class="card-action">
@@ -145,18 +140,23 @@ export default {
   },
   computed: {
     ...mapState(["req", "selected", "selectedCount"]),
-    ...mapGetters(["isListing"]),
+    ...mapGetters(["isListing", "selectedCount"]),
     url() {
       if (!this.isListing) {
         return this.$route.path;
       }
-
-      if (this.selectedCount === 0 || this.selectedCount > 1) {
-        // This shouldn't happen.
-        return;
+      if (this.selectedCount != 1) {
+        // selecting current view image
+        return this.$route.path;
       }
-
       return this.req.items[this.selected[0]].url;
+    },
+    getContext() {
+      let path = this.$route.path.replace("/files/", "./");
+      if (this.selectedCount == 1) {
+        path = path + this.req.items[this.selected[0]].name;
+      }
+      return path;
     },
   },
   async beforeMount() {
@@ -226,9 +226,7 @@ export default {
       return api.getShareURL(share);
     },
     hasDownloadLink() {
-      return (
-        this.selected.length === 1 && !this.req.items[this.selected[0]].isDir
-      );
+      return this.selected.length === 1 && !this.req.items[this.selected[0]].isDir;
     },
     buildDownloadLink(share) {
       return pub_api.getDownloadURL(share);