+++ frontend/src/views/Files.vue
@@ -20,11 +20,10 @@
 import { files as api } from "@/api";
 import { mapState, mapMutations } from "vuex";
 
-import HeaderBar from "@/components/header/HeaderBar";
 import Breadcrumbs from "@/components/Breadcrumbs";
 import Errors from "@/views/Errors";
 import Preview from "@/views/files/Preview.vue";
-import Listing from "@/views/files/Listing.vue";
+import ListingView from "@/views/files/ListingView.vue";
 import Editor from "@/views/files/Editor.vue";
 
 function clean(path) {
@@ -34,11 +33,10 @@ function clean(path) {
 export default {
   name: "files",
   components: {
-    HeaderBar,
     Breadcrumbs,
     Errors,
     Preview,
-    Listing,
+    ListingView,
     Editor,
   },
   data: function () {
@@ -48,18 +46,14 @@ export default {
     };
   },
   computed: {
-    ...mapState(["req", "reload", "loading", "show"]),
+    ...mapState(["req", "reload", "loading"]),
     currentView() {
       if (this.req.type == undefined) {
         return null;
       }
-
       if (this.req.isDir) {
-        return "listing";
-      } else if (
-        this.req.type === "text" ||
-        this.req.type === "textImmutable"
-      ) {
+        return "listingView";
+      } else if (Object.prototype.hasOwnProperty.call(this.req, 'content')) {
         return "editor";
       } else {
         return "preview";
@@ -94,7 +88,7 @@ export default {
     this.setCurrentValue(newView);
   },
   methods: {
-    ...mapMutations(["setLoading","setCurrentView"]),
+    ...mapMutations(["setLoading", "setCurrentView"]),
     async fetchData() {
       // Reset view information.
       this.$store.commit("setReload", false);
@@ -111,7 +105,11 @@ export default {
       if (url[0] !== "/") url = "/" + url;
 
       try {
-        const res = await api.fetch(url);
+        let res = await api.fetch(url);
+        if (!res.isDir) {
+          // get content of file if possible
+          res = await api.fetch(url,true);
+        }
 
         if (clean(res.path) !== clean(`/${this.$route.params.pathMatch}`)) {
           return;