+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. For commit guidelines, please refer to [Standard Version](https://github.com/conventional-changelog/standard-version).
 
+## v0.2.4
+
+- Faature: [create-folder-feature](https://github.com/gtsteffaniak/filebrowser/pull/105)
+- Feature: [playable shared video](https://github.com/filebrowser/filebrowser/issues/2537)
+- Feature: photos, videos, and audio get embedded preview on share instead of icon
+- FIX: sharable link bug, now uses special publicUser
+- Bump go version to 1.22
+- In prep for vue3 migration, npm modules removed:
+  - js-base64
+  - pretty-bytes
+  - whatwg-fetch
+  - lodash.throttle
+  - lodash.clonedeep
+
 ## v0.2.3
 
 - Feature: token expiration time now configurable